fixing

+++ s3.sh
@@ -13,8 +13,21 @@ secretkey=`echo $@ | awk '{print $7}'`
 apiport=`echo $@ | awk '{print $8}'`
 consoleport=`echo $@ | awk '{print $9}'`
 resname=S3-$ipaddr
+minio_image="${MINIO_SERVER_IMAGE:-quay.io/minio/minio:latest}"
+minio_archive="${MINIO_SERVER_ARCHIVE:-/TopStor/minio-server.tar.gz}"
 
 mkdir -p /$pool/$name/$bucket
+
+if ! docker image inspect "$minio_image" >/dev/null 2>&1; then
+ if [ -f "$minio_archive" ]; then
+  docker load -i "$minio_archive" >/dev/null 2>&1 || exit 1
+ fi
+fi
+
+if ! docker image inspect "$minio_image" >/dev/null 2>&1; then
+ exit 1
+fi
+
 docker rm -f $resname
 nmcli conn mod cmynode -ipv4.addresses ${ipaddr}/$ipsubnet
 nmcli conn mod cmynode +ipv4.addresses ${ipaddr}/$ipsubnet
@@ -27,4 +40,4 @@ docker run -d --rm \
   -e MINIO_ROOT_PASSWORD=$secretkey \
   -v /$pool/$name:/data \
   --name $resname \
-  quay.io/minio/minio server /data --console-address ":9001"
+  "$minio_image" server /data --console-address ":9001"
